lesson-4-fix: App moved from data to the root package. Retrofit moved from App to GithubRetrofit.kt
