+++ app/src/main/java/com/rogoz208/mygithubclient/data/retrofit/GithubRetrofit.kt
@@ -1,0 +1,16 @@
+package com.rogoz208.mygithubclient.data.retrofit
+
+import retrofit2.Retrofit
+import retrofit2.adapter.rxjava3.RxJava3CallAdapterFactory
+import retrofit2.converter.gson.GsonConverterFactory
+
+private const val BASE_URL = "https://api.github.com/"
+
+class GithubRetrofit {
+
+    val retrofit: Retrofit = Retrofit.Builder()
+        .baseUrl(BASE_URL)
+        .addConverterFactory(GsonConverterFactory.create())
+        .addCallAdapterFactory(RxJava3CallAdapterFactory.create())
+        .build()
+}
